CAMEL-6476: Introducing StreamCachingStrategy SPI to make it easier to configure and allow 3rd party to plugin custom strategies. Work in progress.

+++ camel-core/src/main/java/org/apache/camel/impl/DefaultStreamCachingStrategy.java
@@ -18,19 +18,52 @@ package org.apache.camel.impl;
 
 import java.io.File;
 
+import org.apache.camel.CamelContext;
+import org.apache.camel.CamelContextAware;
+import org.apache.camel.StreamCache;
 import org.apache.camel.spi.StreamCachingStrategy;
 import org.apache.camel.util.FileUtil;
+import org.apache.camel.util.IOHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class DefaultStreamCachingStrategy extends org.apache.camel.support.ServiceSupport implements StreamCachingStrategy {
+/**
+ * Default implementation of {@link StreamCachingStrategy}
+ */
+public class DefaultStreamCachingStrategy extends org.apache.camel.support.ServiceSupport implements CamelContextAware, StreamCachingStrategy {
 
-    // TODO: Add options to configure more stuff like overflow size et all
     // TODO: Add JMX management
     // TODO: Maybe use #syntax# for default temp dir so ppl can easily configure this
 
+    @Deprecated
+    public static final String THRESHOLD = "CamelCachedOutputStreamThreshold";
+    @Deprecated
+    public static final String BUFFER_SIZE = "CamelCachedOutputStreamBufferSize";
+    @Deprecated
+    public static final String TEMP_DIR = "CamelCachedOutputStreamOutputDirectory";
+    @Deprecated
+    public static final String CIPHER_TRANSFORMATION = "CamelCachedOutputStreamCipherTransformation";
+
     private static final Logger LOG = LoggerFactory.getLogger(DefaultStreamCachingStrategy.class);
+
+    private CamelContext camelContext;
     private File temporaryDirectory;
+    private long spoolThreshold = StreamCache.DEFAULT_SPOOL_THRESHOLD;
+    private String spoolChiper;
+    private int bufferSize = IOHelper.DEFAULT_BUFFER_SIZE;
+    private boolean removeTemporaryDirectoryWhenStopping = true;
+
+    public CamelContext getCamelContext() {
+        return camelContext;
+    }
+
+    public void setCamelContext(CamelContext camelContext) {
+        this.camelContext = camelContext;
+    }
+
+    public void setTemporaryDirectory(String path) {
+        this.temporaryDirectory = new File(path);
+    }
 
     public void setTemporaryDirectory(File path) {
         this.temporaryDirectory = path;
@@ -40,8 +73,60 @@ public class DefaultStreamCachingStrategy extends org.apache.camel.support.Servi
         return temporaryDirectory;
     }
 
+    public long getSpoolThreshold() {
+        return spoolThreshold;
+    }
+
+    public void setSpoolThreshold(long spoolThreshold) {
+        this.spoolThreshold = spoolThreshold;
+    }
+
+    public String getSpoolChiper() {
+        return spoolChiper;
+    }
+
+    public void setSpoolChiper(String spoolChiper) {
+        this.spoolChiper = spoolChiper;
+    }
+
+    public int getBufferSize() {
+        return bufferSize;
+    }
+
+    public void setBufferSize(int bufferSize) {
+        this.bufferSize = bufferSize;
+    }
+
+    public boolean isRemoveTemporaryDirectoryWhenStopping() {
+        return removeTemporaryDirectoryWhenStopping;
+    }
+
+    public void setRemoveTemporaryDirectoryWhenStopping(boolean removeTemporaryDirectoryWhenStopping) {
+        this.removeTemporaryDirectoryWhenStopping = removeTemporaryDirectoryWhenStopping;
+    }
+
     @Override
     protected void doStart() throws Exception {
+        String bufferSize = camelContext.getProperty(BUFFER_SIZE);
+        String hold = camelContext.getProperty(THRESHOLD);
+        String chiper = camelContext.getProperty(CIPHER_TRANSFORMATION);
+        String dir = camelContext.getProperty(TEMP_DIR);
+
+        if (bufferSize != null) {
+            this.bufferSize = camelContext.getTypeConverter().convertTo(Integer.class, bufferSize);
+        }
+        if (hold != null) {
+            this.spoolThreshold = camelContext.getTypeConverter().convertTo(Long.class, hold);
+        }
+        if (chiper != null) {
+            this.spoolChiper = chiper;
+        }
+        if (dir != null) {
+            this.temporaryDirectory = camelContext.getTypeConverter().convertTo(File.class, dir);
+        }
+
+        LOG.info("StreamCaching in use with {}", this.toString());
+
         // create random temporary directory if none has been created
         if (temporaryDirectory == null) {
             temporaryDirectory = FileUtil.createNewTempDir();
@@ -60,9 +145,18 @@ public class DefaultStreamCachingStrategy extends org.apache.camel.support.Servi
 
     @Override
     protected void doStop() throws Exception {
-        if (temporaryDirectory != null) {
+        if (temporaryDirectory != null  && isRemoveTemporaryDirectoryWhenStopping()) {
             LOG.info("Removing temporary directory {}", temporaryDirectory);
             FileUtil.removeDir(temporaryDirectory);
         }
+    }
+
+    @Override
+    public String toString() {
+        return "DefaultStreamCachingStrategy["
+            + "temporaryDirectory=" + temporaryDirectory
+            + ", spoolThreshold=" + spoolThreshold
+            + ", spoolChiper=" + spoolChiper
+            + ", bufferSize=" + bufferSize + "]";
     }
 }
